requirements.txt
# genral modules

asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
psycopg2-binary==2.9.4
sqlparse==0.4.3
python-decouple
pillow
djangorestframework 

# third party  modules
markdown
django-filter
drf-yasg
pytest
pytest-django